create vao for waypoint path remove conway stuff draw lines

+++ dto/activity.go
@@ -2,5 +2,5 @@ package dto
 
 type Activity struct {
 	ActivityType string  `json:"activityType,omitempty"`
-	Probability  float64 `json:"probability"`
+	Probability  float32 `json:"probability"`
 }
