+++ dto/location.go
@@ -8,6 +8,6 @@ type Location struct {
 	Address               string     `json:"address,omitempty"`
 	Name                  string     `json:"name,omitempty"`
 	LocationConfidence    int        `json:"placeConfidence,omitempty"`
-	CalibratedProbability float64    `json:"calibratedProbability,omitempty"`
+	CalibratedProbability float32    `json:"calibratedProbability,omitempty"`
 	AccuracyMetres        int        `json:"accuracyMeters,omitempty"`
 }
