+++ dto/waypointpath.go
@@ -4,7 +4,7 @@ type WaypointPath struct {
 	Waypoints      []Point `json:"waypoints,omitempty"`
 	Source         string  `json:"source,omitempty"`
 	RoadSegment    []Place `json:"roadSegment,omitempty"`
-	DistanceMetres float64 `json:"distanceMeters,omitempty"`
+	DistanceMetres float32 `json:"distanceMeters,omitempty"`
 	TravelMode     string  `json:"travelMode,omitempty"`
-	Confidence     float64 `json:"confidence,omitempty"`
+	Confidence     float32 `json:"confidence,omitempty"`
 }
